vivita-inventory/requirements.txt
streamlit==1.29.0
supabase==2.0.3
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
python-jose==3.3.0
requests==2.31.0
typing-extensions==4.9.0
pydantic==2.5.3
numpy>=1.26.3
